requirements.txt
Django==1.11.12
requests==2.18.1
kombu==3.0.37
pytz==2017.3
celery==3.1.25
django_celery==3.2.2
djangorestframework==3.8.2
simplejson==3.15.0
